Added a StorageIndex for the source storage to reduce LIST calls
Before every restructuring or cleaning operation, the index is updated.
The StorageIndex can be updated with a configurable sync time to make a
full sync, otherwise it just updates directories that have files in them.
A separate sync time can be set to also scan empty directories.
The first implementation is only a InMemoryStorageIndex. For
very large datasets, a file-based index might be needed.